added a shorthand php echo statement section to class1.md
In PHP, understanding the shorthand echo tag is essential. Although it might not have been discussed in class, not having clarity on this topic could lead to confusion for many. Therefore, after the PHP echo statement, I have included an explanation. This way, students won't face difficulties if they encounter the use of this shorthand tag in a class in the future.